Four Hundred Seventy Three Commit: Refactor findRootCategories() method

+++ ShopmeWebParent/ShopmeBackEnd/src/main/java/com/shopme/admin/category/CategoryPageInfo.java
@@ -2,7 +2,7 @@ package com.shopme.admin.category;
 
 public class CategoryPageInfo {
 	private int totalPages;
-	private int totalElements;
+	private long totalElements;
 	
 	public int getTotalPages() {
 		return totalPages;
@@ -10,10 +10,10 @@ public class CategoryPageInfo {
 	public void setTotalPages(int totalPages) {
 		this.totalPages = totalPages;
 	}
-	public int getTotalElements() {
+	public long getTotalElements() {
 		return totalElements;
 	}
-	public void setTotalElements(int totalElements) {
+	public void setTotalElements(long totalElements) {
 		this.totalElements = totalElements;
 	}
